# internal/httpserve/handlers/login.go
package handlers

import (
	"fmt"
	"net/http"
	"strings"

	echo "github.com/theopenlane/echox"

	"github.com/theopenlane/utils/rout"

	"github.com/theopenlane/iam/auth"

	"github.com/theopenlane/utils/passwd"

	"github.com/theopenlane/core/common/enums"
	models "github.com/theopenlane/core/common/openapi"
	"github.com/theopenlane/core/internal/ent/generated/privacy"
	"github.com/theopenlane/core/pkg/logx"
	"github.com/theopenlane/core/pkg/metrics"
	sso "github.com/theopenlane/core/pkg/ssoutils"
)

// LoginHandler validates the user credentials and returns a valid cookie
// this handler only supports username password login
func (h *Handler) LoginHandler(ctx echo.Context) error {
	req, err := BindAndValidate[models.LoginRequest](ctx)
	if err != nil {
		metrics.RecordLogin(false)
		logx.FromContext(ctx.Request().Context()).Warn().Err(err).Msg("invalid login request")

		return h.InvalidInput(ctx, err)
	}

	reqCtx := ctx.Request().Context()

	// the Openlane support identity is virtual and authenticated entirely from configuration; detect it
	// before any database lookup so this email can never authenticate against a database user, and route
	// it to the configuration backed support login (password factor plus second factor identity provider)
	if h.SupportAccessConfig.Enabled && strings.EqualFold(req.Username, h.SupportAccessConfig.Email) {
		return h.supportFirstFactor(ctx, req)
	}

	// check user in the database, username == email and ensure only one record is returned
	user, err := h.getUserByEmail(reqCtx, req.Username)
	if err != nil {
		metrics.RecordLogin(false)
		logx.FromContext(reqCtx).Info().Str("email", req.Username).Err(err).Msg("unable to find user by email")

		return h.BadRequest(ctx, ErrLoginFailed)
	}

	if user.Edges.Setting.Status != enums.UserStatusActive {
		metrics.RecordLogin(false)
		logx.FromContext(reqCtx).Info().Str("email", req.Username).Msg("user not active")

		return h.BadRequest(ctx, ErrLoginFailed)
	}

	allowCtx := privacy.DecisionContext(reqCtx, privacy.Allow)

	orgStatus := h.orgEnforcementsForUser(allowCtx, req.Username)
	if orgStatus != nil && orgStatus.Enforced {
		return ctx.Redirect(http.StatusFound, sso.SSOLogin(ctx.Echo(), orgStatus.OrganizationID))
	}

	if user.Password == nil {
		metrics.RecordLogin(false)
		logx.FromContext(reqCtx).Info().Str("email", req.Username).Msg("no password set for user")

		return h.BadRequest(ctx, ErrLoginFailed)
	}

	// verify the password is correct
	valid, err := passwd.VerifyDerivedKey(*user.Password, req.Password)
	if err != nil || !valid {
		metrics.RecordLogin(false)
		logx.FromContext(reqCtx).Info().Str("email", req.Username).Msg("invalid password provided during login")

		return h.BadRequest(ctx, ErrLoginFailed)
	}

	if !user.Edges.Setting.EmailConfirmed {
		metrics.RecordLogin(false)
		logx.FromContext(reqCtx).Info().Str("email", req.Username).Msg("user email not verified, unable to login")

		return h.BadRequest(ctx, fmt.Errorf("%w: please check your email and verify your account", auth.ErrUnverifiedUser))
	}

	// set context for remaining request based on logged in user
	userCtx := setAuthenticatedContext(reqCtx, user)

	// create new claims for the user
	auth, err := h.AuthManager.GenerateUserAuthSession(userCtx, ctx.Response().Writer, user)
	if err != nil {
		logx.FromContext(reqCtx).Error().Err(err).Msg("unable to create new auth session")

		return h.InternalServerError(ctx, ErrProcessingRequest)
	}

	if err := h.updateUserLastSeen(userCtx, user.ID, enums.AuthProviderCredentials); err != nil {
		logx.FromContext(reqCtx).Error().Err(err).Msg("unable to update last seen")

		return h.InternalServerError(ctx, ErrProcessingRequest)
	}

	// check if orgStatus is enforced, but user has not yet configured
	// if not yet configured we want to direct to the setup first
	tfaSetupRequired := false

	if orgStatus != nil && orgStatus.OrgTFAEnforced {
		// Check if user has TFA enabled
		if user.Edges.Setting == nil || !user.Edges.Setting.IsTfaEnabled {
			tfaSetupRequired = true
		}
	}

	out := models.LoginReply{
		Reply:            rout.Reply{Success: true},
		TFAEnabled:       user.Edges.Setting.IsTfaEnabled,
		TFASetupRequired: tfaSetupRequired,
		Message:          "success",
		AuthData:         *auth,
	}

	metrics.RecordLogin(true)

	return h.Success(ctx, out)
}