# internal/httpserve/specs/openlane.openapi.json
{
  "components": {
...
    "schemas": {
...
      "ErrorReply": {
        "properties": {
          "error": {
            "description": "The error message if the request was unsuccessful",
            "type": "string"
          },
          "error_code": {
            "description": "The error code if the request was unsuccessful",
            "type": "string"
          },
          "success": {
            "description": "Whether or not the request was successful or not",
            "type": "boolean"
          },
          "unverified": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "InviteReply": {
        "properties": {
          "access_token": {
            "description": "The access token to be used for authentication",
            "type": "string"
          },
          "email": {
            "description": "The email address of the user",
            "example": "jsnow@example.com",
            "type": "string"
          },
          "error": {
            "description": "The error message if the request was unsuccessful",
            "type": "string"
          },
          "error_code": {
            "description": "The error code if the request was unsuccessful",
            "type": "string"
          },
          "joined_org_id": {
            "description": "The ID of the organization the user joined",
            "example": "01JJFVMGENQS9ZG3GVA50QVX5E",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "needs_sso": {
            "type": "boolean"
          },
          "refresh_token": {
            "description": "The refresh token to be used to refresh the access token after it expires",
            "type": "string"
          },
          "role": {
            "description": "The role the user has in the organization",
            "example": "admin",
            "type": "string"
          },
          "session": {
            "description": "The short-lived session token required for authentication",
            "type": "string"
          },
          "success": {
            "description": "Whether or not the request was successful or not",
            "type": "boolean"
          },
          "token_type": {
            "description": "The type of token being returned",
            "example": "bearer",
            "type": "string"
          },
          "unverified": {
            "type": "boolean"
          },
          "user_id": {
            "description": "The ID of the user that was created",
            "example": "01J4EXD5MM60CX4YNYN0DEE3Y1",
            "type": "string"
          }
        },
        "type": "object"
      },
      "LoginReply": {
        "properties": {
          "access_token": {
            "description": "The access token to be used for authentication",
            "type": "string"
          },
          "error": {
            "description": "The error message if the request was unsuccessful",
            "type": "string"
          },
          "error_code": {
            "description": "The error code if the request was unsuccessful",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "refresh_token": {
            "description": "The refresh token to be used to refresh the access token after it expires",
            "type": "string"
          },
          "session": {
            "description": "The short-lived session token required for authentication",
            "type": "string"
          },
          "success": {
            "description": "Whether or not the request was successful or not",
            "type": "boolean"
          },
          "tfa_enabled": {
            "type": "boolean"
          },
          "tfa_required": {
            "type": "boolean"
          },
          "token_type": {
            "description": "The type of token being returned",
            "example": "bearer",
            "type": "string"
          },
          "unverified": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "password": {
            "description": "The password associated with the account",
            "example": "Wint3rIsC0ming123!",
            "type": "string"
          },
          "reason": {
            "description": "For Openlane support login only: the reason for accessing the organization",
            "type": "string"
          },
          "target_organization_id": {
            "description": "For Openlane support login only: the organization to access",
            "type": "string"
          },
          "username": {
            "description": "The email address associated with the existing account",
            "example": "jsnow@example.com",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Bearer token authentication using API tokens.\nThe token identifies the organization and grants permissions to manage\nusers and groups within that organization's scope.\n",
        "scheme": "bearer",
        "type": "http"
      },
      "apiKey": {
        "description": "API Key authentication, the key must be a valid API key",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Bearer authentication, the token must be a valid API token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
...
  "info": {
    "contact": {
      "email": "support@theopenlane.io",
      "name": "Openlane",
      "url": "https://www.theopenlane.io"
    },
    "description": "Openlane's API services are designed to provide a simple and easy to use interface for interacting with the Openlane platform. This API is designed to be used by both internal and external clients to interact with the Openlane platform.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Openlane OpenAPI 3.1.1 Specifications",
    "version": "v0.0.0-dev"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/v1/invite": {
      "get": {
        "description": "Accept an organization invitation",
        "operationId": "OrganizationInviteAccept",
        "parameters": [
          {
            "description": "The token to be used to accept the invitation, token is sent via email",
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "InviteResponse": {
                    "value": {
                      "access_token": "access_token",
                      "email": "",
                      "joined_org_id": "1234",
                      "message": "Welcome to your new organization!",
                      "refresh_token": "refresh_token",
                      "role": "admin",
                      "session": "session",
                      "success": true,
                      "token_type": "bearer",
                      "user_id": "1234"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/InviteReply"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "bad request",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "unauthorized",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "internal server error",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "OrganizationInviteAccept",
        "tags": [
          "Invites"
        ]
      }
    },
    "/v1/login": {
      "post": {
        "description": "Login is oriented towards human users who use their email and password for authentication. Login verifies the password submitted for the user is correct by looking up the user by email and using the argon2 derived key verification process to confirm the password matches. Upon authentication an access token and a refresh token with the authorized claims of the user are returned. The user can use the access token to authenticate to our systems. The access token has an expiration and the refresh token can be used with the refresh endpoint to get a new access token without the user having to log in again. The refresh token overlaps with the access token to provide a seamless authentication experience and the user can refresh their access token so long as the refresh token is valid",
        "operationId": "LoginHandler",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "AvailableAuthTypeRequest": {
                  "value": {
                    "password": "",
                    "username": "sfunky@theopenlane.io"
                  }
                },
                "LoginSuccessRequest": {
                  "value": {
                    "password": "mitb!",
                    "username": "sfunky@theopenlane.io"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "description": "Request body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "LoginSuccessResponse": {
                    "value": {
                      "access_token": "access_token",
                      "message": "",
                      "refresh_token": "refresh_token",
                      "session": "session",
                      "success": true,
                      "tfa_enabled": true,
                      "token_type": "bearer"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/LoginReply"
                }
              }
            },
            "description": "OK"
          },
          "302": {
            "description": "Found",
            "headers": {
              "Location": {
                "description": "Destination URL of the redirect",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "bad request",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "unauthorized",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "forbidden",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "not found",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "too many requests",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "internal server error",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [],
        "summary": "Login",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Openlane API Server",
      "url": "https://api.theopenlane.io"
    }
  ],
...
}